Fix checks using double/adjacent clue filtering regex
We need to allow for 4-char words of the following sort
  (1c)(1c)(2c)  \_  Single-char clue followed by a 3-char double clue
  (1c)(2c)(1c)  /
  (2c)(1c)(1c)  -- Double-char clue followed by any 2-char double clue
where (1c) represents any 1-char clue and (2c) any 2-char clue.
We assume for this there is exactly one 2-char clue available and the
rest are 1-char clues.

We also need to allow for other words made up of an odd number of clues,
which I accomodate by only checking for prefixes of two double-clues (in
addition to the above 4-char word check).