add argument parsing
Add parsing of args to choose grid file, words file, and/or to choose to
select these randomly instead of prompting user. Random selection only
occurs for files not otherwise selected on the commandline. Duplicate
arguments are not checked for, so it is the last version of each arg
that gets used. Unsupported arg and positional args throw an error.