Keep track of prefix paths to shortcut searches
As we search for paths for a word and discover the possible paths for
its prefix strings, maintain a map from prefix strings to possible paths
so that when investigate later words we can short-circuit / start at the
longest prefix paths that has already been handled. For example, if we
search for 'condition' and also for 'conduit', then the search for
'condition' will take us through 'c', 'co', etc, until we either fail to
find a path for a "prefix" of 'condition' or success. We can start our
search for 'conduit' based on the longest prefix that already been
searched, 'condui', 'condu', 'cond', 'con', 'co', or 'c'.

Not that doing this make it important for us to search for all possible
paths for any prefix, and so we can't short-circuit successes on the
first successful path found (or if we do, then we need to not record the
prefix paths for the current word).