rm attempt to handle 'impossible prefixes'
It is still not correctly implemented at this point and I determined
that it would be too complex (relative to the natural flow of the rest
of the code) to correctly implement the feature to detect prefix strings
that can't be extended. For example, saw we have a grid such that
  'cond'   - has paths
  'condom' - doesn't have paths
and now we want to check 'condominium' for paths. Well, even though
'condom' doesn't have paths, if there is a multi-char clue containing
'mi' then we might be able to build 'condomi' using that clue, and thus
might be able to build 'condominium'.

While I /could/ attempt to deal with such situations, it would be a
notable enough increase in complexity (away from what I condsider to be
the natural/intuitive flow of the code) that I don't want to attempt it
right now.