Add ability to togger coloring on/off
It's not perfect. There are a few lines that get colored even if the
user provides '--no-color' arg. This is b/c of the setup order for
logging, coloration (default on), and then arg parsing, with embedded
log statements in some setup functions.

This change allows me to change some 'echo' statements during arg
parsing into true 'logError' statements.