Add a testing feature
The user can pass in a -t or --test option with a provided md5sum-style
hash (32-char lowercase hex string) that will be used to test the final
words list file. This should make it easy for me to add a test.sh script
to easily validate if code changes have changed the final filtering for
any combo of grid and words list files I have in the codebase.