rm failed attempt at keep track of prefix paths
When I added test cases and checked the current filtered results against
past filtered results I discovered that the attempt to shorten path
searches by keeping track of previously computed paths for any prefix
was causing some paths to be missed. The paths that were missed involved
multi-char clues. For example, from sample grid file 02 and the
internation words list: the prefix 'shie' is found but the paths for
'shie' cannot be extended to 'shier' and so that word was being dropped.
Yet there IS a path for 'shier' - a path that uses the 2-char clue 'er'.

There is probably an elegant way to patch* the logic without making it
too convoluted, but right now I will instead test out adding
short-circuiting logic during path construction so that for each word we
stop looking once the first path is found; with the removal of keeping
track of prefixToPath map, we don't need to find all such paths.

* Hypothesis: keep track of the length of the longest clue and as well
as a mapping from clue-length (1 char, 2 chars, etc) to positions;
update the logic for function extractPathsForLongestPrefix so that
once the first/longest previously prefix has been found (if any), we
test out trimming char(s) off the selected prefix and adding any clues
longer than the number of chars trimmed off the prefix.