Perform final path-search filter in reverse order
And keep track of the strings for which paths have already been
encountered, without bothering to keep track of the paths themselves.
This way, since we are searching in reverse lexicographical order, any
words that are prefixes of other words will occur later and we can easy
check if a path has been encountered before. If yes, we can just add
that word to our final words list rather than searching for a specific
path.